It is Project Plan and Description share application to a project fair or competition
The main goal of this project is to share project plan and description to a university project competition or a project fair. In that application, there have two panel one is for administrator or teacher who will see and handle your project plan and everything and another panel is for students or participant panel where a student/PARTICIPANT will register then the can provide their project and description to project fair.